split up the BulkSelectionPanel into multiple components
