+++ frontend/src/components/pageEditor/bulkSelectionPanel/BulkSelection.ts
