+++ frontend/src/components/pageEditor/bulkSelectionPanel/BulkSelectionPanel.module.css
